requirements.txt
ultralytics
opencv-python
numpy
pyserial
scikit-learn